Move all files to top level directory
All of Auctora will live in an App Engine application. Statics and
HTML will be served by our App Engine application backend. For this
reason, the App Engine code and configuration will be at the top level.

We can figure out folders for different parts of the project later, like
the tablet/desktop site and the phone site.