Host Auctora pitch slides on the application.
The slides we used for our pitches are a website created using
simplePrez (http://github.com/eric-zeng/simplePrez). Move the
presentation files to this project, and modify hello.go to serve
the content. The slides are available at
tidy-nomad-842.appspot.com/slides

Hopefully this is a good example for how to write HTTP handlers and
serve files from the backend.